feat: Frameworks gap closure — Sprints 23-25 (Run Payments, POS, AP)
Sprint 23: Run Payments Integration
- PaymentGateway interface for pluggable payment providers
- Run Payments HTTP client (charge/capture/void/refund)
- RunPaymentsForm.tsx with Runner.js PCI-compliant tokenization
- New endpoints: POST /api/payments/intent, /card, /refund
- Migration 026: gateway fields, payment_refunds, tender_lines
- Config: RUN_PAYMENTS_API_KEY, PUBLIC_KEY, BASE_URL, ENV

Sprint 24: Retail POS Module
- Full backend/internal/pos/ package (model/repo/service/handler)
- 8 API routes: transaction lifecycle, product search
- POSTerminal.tsx full-screen counter sales UI
- Inventory auto-deduct on completion, restore on void
- Migration 027: pos_transactions, line_items, tenders, registers

Sprint 25: Accounts Payable
- Full backend/internal/ap/ package (model/repo/service/handler)
- 7 API routes: vendor invoices, approval, payments, aging
- Multi-invoice payment application within transactions
- AP aging report (current/30/60/90+ buckets)
- Migration 028: vendor_invoices, invoice_lines, ap_payments

Also includes: gap analysis docs, closure roadmap (Sprints 23-34)

+++ backend/internal/portal/model.go
@@ -76,9 +76,9 @@ type PortalInvoiceDTO struct {
 	ID           uuid.UUID       `json:"id"`
 	OrderID      uuid.UUID       `json:"order_id"`
 	Status       string          `json:"status"`
-	TotalAmount  int64           `json:"total_amount"`
-	Subtotal     int64           `json:"subtotal"`
-	TaxAmount    int64           `json:"tax_amount"`
+	TotalAmount  float64         `json:"total_amount"`
+	Subtotal     float64         `json:"subtotal"`
+	TaxAmount    float64         `json:"tax_amount"`
 	PaymentTerms string          `json:"payment_terms"`
 	DueDate      *time.Time      `json:"due_date"`
 	PaidAt       *time.Time      `json:"paid_at"`
